Renamed generate_data to functions_for_data_generation.
