+++ src/paper/research_paper.tex
@@ -657,7 +657,7 @@ Our approach to the problem is cross-validation on a two dimensional grid. Due t
 \section{Code}
 
 \lstset{title=Data generating functions}
-\pythonexternal{../../../src/data_management/generate_data.py}
+\pythonexternal{../../../src/data_management/functions_for_data_generation.py}
 
 \lstset{title=Functions which solve the optimization problem}
 \pythonexternal{../../../src/model_code/fused_lasso_primal.py}
